requirements.txt
websockets
psycopg2-binary
python-dotenv
Flask
pyjwt